YAN-583: Solution Interval Feature

See merge request ska-telescope/icrar-leap-accelerate!20

+++ src/icrar/leap-accelerate/algorithm/cpu/PhaseMatrixFunction.cc
@@ -66,7 +66,8 @@ namespace cpu
             if(a1(n) != a2(n))
             {
                 // skip entry if data not flagged
-                if(!fg(n) && ((!refAnt.is_initialized()) || ((refAnt.is_initialized()) && ((a1(n) == refAnt.get()) || (a2(n) == refAnt.get())))))
+                if(!fg(n) && ((!refAnt.is_initialized())
+                || ((refAnt.is_initialized()) && ((static_cast<unsigned int>(a1(n)) == refAnt.get()) || (static_cast<unsigned int>(a2(n)) == refAnt.get())))))
                 {
                     A(k, a1(n)) = 1;
                     A(k, a2(n)) = -1;
